feat: add crud comment function in useComment
