+++ src/feature/comment/ui/CommentEditDialog.tsx
@@ -1,38 +1,11 @@
 import { Button, Dialog, Textarea } from "../../../shared/ui"
 import { useCommentStore } from "../model/store"
 import { Comment } from "../../../entities"
-import { requestApi } from "../../../shared/lib"
-import { UpsertComment } from "../type"
+import { useComment } from "../model/hook"
 
 export const CommentEditDialog = () => {
-  const {
-    comments,
-    setComments,
-    selectedComment,
-    setSelectedComment,
-    showEditCommentDialog,
-    setShowEditCommentDialog,
-  } = useCommentStore()
-
-  // 댓글 업데이트
-  const updateComment = async () => {
-    try {
-      const { result, data } = await requestApi<UpsertComment>(`/api/comments/${selectedComment?.id}`, {
-        method: "PUT",
-        body: JSON.stringify({ body: selectedComment?.body }),
-      })
-
-      if (result && data) {
-        setComments({
-          ...comments,
-          [data.postId]: (comments[data.postId] ?? []).map((c) => (c.id === data.id ? data : c)),
-        })
-        setShowEditCommentDialog(false)
-      }
-    } catch (error) {
-      console.error("댓글 업데이트 오류:", error)
-    }
-  }
+  const { selectedComment, setSelectedComment, showEditCommentDialog, setShowEditCommentDialog } = useCommentStore()
+  const { updateComment } = useComment()
 
   return (
     <Dialog open={showEditCommentDialog} handleChange={setShowEditCommentDialog} title="댓글 수정">
